replace ITileEntityProvider with annotation based system
This allows us to use Block#hasTileEntity() and Block#createTileEntity()
which is now considered to be *best practice*.

TileEntity registration is done via an annotation. If we need to support
multiple tile entities per block later on, we can use the
HasTileEntity annotation multiple times but we will need to add support
for the container annotation `HasTileEntities` otherwise they wont get
registered.