make leylight bore rotate to a direction point
The bore can now be rightclicked (will require a special item later on),
which allows setting the mining direction of the bore. The rotation
will be completly clientside, the server just has progress which is set
to 1 and will be decreased each tick until it reaches 0. At this point
the server assumes the movement is done (and sends a package) and we
can start the mining logic.